Вынес классы карты и колоды в отдельный модуль
